refactor: rename to `@nuxt/schema/build-env`
