+++ packages/schema/src/build-env.ts
@@ -1,3 +1,3 @@
-import './types/env'
+import './types/build-env'
 
 export const builders = ['vite', 'webpack']
