+++ packages/schema/src/types/build-env/vite.ts
